+++ src/tcp.h
@@ -12,7 +12,7 @@
  * @param port The port to listen on.
  * @return The server socket file descriptor.
  */
-i32 start_server(u16 port);
+int start_server(int port);
 
 /**
  * Polls for a new connection on the given server socket.
@@ -21,4 +21,4 @@ i32 start_server(u16 port);
  * @return An async result indicating whether the poll is ready or pending, and the client socket
  * file descriptor if the poll is ready.
  */
-async_result_t poll_accept_connection(i32 server_fd);
+async_result_t poll_accept_connection(int server_fd);
